// projects/ngxsmk-datepicker/src/lib/interfaces/datepicker-classes.interface.ts
export interface DatepickerClasses {
  wrapper?: string;
  inputGroup?: string;
  input?: string;
  clearBtn?: string;
  calendarBtn?: string;
  popover?: string;
  container?: string;
  calendar?: string;
  header?: string;
  navPrev?: string;
  navNext?: string;
  dayCell?: string;
  footer?: string;
  closeBtn?: string;
}